+++ CHANGES.md
@@ -18,3 +18,4 @@
 - Improve software tests and documentation
 - Settings: Obtain environment variables from `.env` file
 - Settings: Obtain job store schema- and table names alongside database address
+- Packaging / Refactoring
