--- requirements-dev.txt
@@ -1,15 +1,0 @@
-# Formatter and linter
-black<24
-mypy==1.7.1
-poethepoet<0.25
-pyproject-fmt<1.6
-ruff==0.1.6
-validate-pyproject<0.16
-
-# Software tests
-cratedb-toolkit==0.0.2
-httpx<0.26
-psycopg2-binary<3
-pytest<8
-pytest-cov<5
-pytest-mock<4
