--- requirements.txt
@@ -1,16 +1,0 @@
-APScheduler<4
-click<9
-colorama<0.5
-colorlog<7
-crate[sqlalchemy]
-fastapi<0.105
-halo<0.1
-icecream<3
-Jinja2<4
-MarkupSafe<3
-pueblo[fileio] @ git+https://github.com/pyveci/pueblo@more-io
-pydantic<3
-python-dotenv[cli]<2
-python-multipart==0.0.6
-uvicorn<0.25
-watchdog<4
