+++ tests/test_cli.py
@@ -3,12 +3,12 @@ from unittest.mock import MagicMock
 import pytest
 from click.testing import CliRunner
 
-from cli import cli
+from supertask.cli import cli
 
 
 @pytest.fixture
 def st_wait_noop(mocker):
-    mocker.patch("core.Supertask.wait")
+    mocker.patch("supertask.core.Supertask.wait")
 
 
 def test_cli_help(st_wait_noop):
@@ -52,7 +52,7 @@ def test_cli_storage_memory(st_wait_noop):
 def test_cli_http_service(mocker, st_wait_noop):
     runner = CliRunner(env={"ST_STORE_ADDRESS": "memory://", "ST_HTTP_LISTEN_ADDRESS": "localhost:3333"})
 
-    start_http_service_mock: MagicMock = mocker.patch("core.Supertask.start_http_service")
+    start_http_service_mock: MagicMock = mocker.patch("supertask.core.Supertask.start_http_service")
     result = runner.invoke(
         cli,
         args="--verbose",
