+++ tests/test_job.py
@@ -1,4 +1,4 @@
-from jobs import my_job
+from supertask.vendor.jobs import my_job
 
 
 def test_my_job(mocker, capsys):
